rename p2p/discovery/backoff package, run go mod tidy
