+++ p2p/discovery/backoff/backoffcache.go
@@ -1,4 +1,4 @@
-package discovery
+package backoff
 
 import (
 	"context"
