+++ p2p/discovery/backoff/backoffconnector.go
@@ -1,10 +1,11 @@
-package discovery
+package backoff
 
 import (
 	"context"
-	lru "github.com/hashicorp/golang-lru"
 	"sync"
 	"time"
+
+	lru "github.com/hashicorp/golang-lru"
 
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
