+++ p2p/discovery/backoff/mocks_test.go
@@ -1,4 +1,4 @@
-package discovery
+package backoff
 
 import (
 	"context"
